feat: VM 核心编译过裸机 MCU(thumbv7em)—— M5.2 依赖手术完成,crate graph 全程无 std
依赖手术(此前 no_std 构建靠依赖把 std 连进 graph,f64 inherent 方法等
因此糊弄过关;本次彻底):
- 删 5 个死依赖:once_cell/chrono/sha2/rand/tracing(M0/M5.4 清理后遗留);
  tempfile 移 dev-dependencies(仅测试用)
- anyhow/serde/serde_json 切 no_std 模式(default-features=false+alloc,
  std feature 重新点亮 */std)
- toml/serde_yaml/dashmap optional 绑 std:val/de.rs YAML/TOML 解码 gate
  (no_std 保 JSON);compat 加 SharedMap(std=DashMap,no_std=spin Mutex
  +hashbrown,Arc 存储免 V:Clone,双态同 get/insert→.value() 表面)
- compat::float(fract/abs):f64 inherent 方法在 std,no_std 走 libm
- AtomicU64→AtomicUsize(Cortex-M 无 64 位原子);crate-type 去 staticlib
  (无消费者,no_std 下要求 allocator/panic handler)
CI thumbv7em 步骤加 lk-core 全 VM 核心构建守卫。
验证:workspace 106 套测试全 ok / host no_std / wasm32 / clippy 全绿;
std 路径依赖语义不变(bench 不受影响:热路径零代码变化)。

+++ core/src/macro_system/procedural.rs
@@ -17,7 +17,10 @@ use crate::{
     token::{ParseError, Position, Span, Token, Tokenizer},
 };
 use core::fmt;
-use core::sync::atomic::{AtomicU64, Ordering};
+// Temp-file name uniqueness counter: 64-bit atomics do not exist on
+// Cortex-M-class targets (`target_has_atomic = "64"` is off), and usize is
+// plenty for a per-process temp counter.
+use core::sync::atomic::{AtomicUsize, Ordering};
 use core::time::Duration;
 use serde::{Deserialize, Serialize};
 // Running an external proc-macro provider needs a process + filesystem, which
@@ -32,7 +35,7 @@ use std::{
 
 pub const PROC_MACRO_PROTOCOL_VERSION: u32 = 1;
 const BUILTIN_SHOW_TRAIT: &str = "__LKShow";
-static PROC_MACRO_TEMP_COUNTER: AtomicU64 = AtomicU64::new(0);
+static PROC_MACRO_TEMP_COUNTER: AtomicUsize = AtomicUsize::new(0);
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
 pub enum ProcMacroKind {
